Bootstrap Percentile Confidence Intervals

Bootstrap Percentile Confidence Intervals

Usage

BootCI(x, alpha = 0.05)

Arguments

x

Numeric matrix. Output of PBootVAROLS(), PBootVARLasso(), RBootVAROLS(), or RBootVARLasso().

alpha

Numeric. Significance level.

Value

A list with two elements, namely ll for the lower limit and ul for the upper limit.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

set.seed(42)
# Parametric bootstrap
system.time(
  pb <- PBootVAROLS(
    data = dat_p2,
    p = 2,
    B = 10,
    burn_in = 20
  )
)
#>    user  system elapsed 
#>   0.027   0.004   0.017 
pb$est
#>           [,1]        [,2]        [,3]        [,4]        [,5]         [,6]
#> [1,] 0.7899726  0.36836768 0.008539283  0.02231498  0.11338780  0.024505442
#> [2,] 1.0001995  0.01334496 0.481949237 -0.01730029 -0.03223315  0.234373552
#> [3,] 1.0666814 -0.03022951 0.014945306  0.60741736 -0.01610408 -0.006149189
#>             [,7]
#> [1,] -0.01143384
#> [2,]  0.01815779
#> [3,]  0.29620975
BootCI(pb)
#> $ll
#>           [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> [1,] 0.3838074  0.30829901 -0.01704828 -0.03057283  0.04649171 -0.02836141
#> [2,] 0.6478630 -0.02957636  0.45446362 -0.08252004 -0.06703509  0.19405729
#> [3,] 0.8696531 -0.07535331 -0.02489120  0.55981382 -0.03983556 -0.05304907
#>              [,7]
#> [1,] -0.061178743
#> [2,] -0.003047983
#> [3,]  0.271767462
#> 
#> $ul
#>          [,1]        [,2]       [,3]       [,4]       [,5]       [,6]
#> [1,] 1.227579  0.39076365 0.07679172 0.06472094 0.15950531 0.05946045
#> [2,] 1.132903  0.02884185 0.52846257 0.02128911 0.04365536 0.29238903
#> [3,] 1.493904 -0.02126475 0.07561374 0.63468576 0.06009620 0.00894426
#>            [,7]
#> [1,] 0.04043717
#> [2,] 0.08005482
#> [3,] 0.32298319
#> 
system.time(
  pb <- PBootVARLasso(
    data = dat_p2,
    p = 2,
    B = 10,
    burn_in = 20,
    n_lambdas = 100,
    crit = "ebic",
    max_iter = 1000,
    tol = 1e-5
  )
)
#>    user  system elapsed 
#>  20.831   4.466  12.712 
pb$est
#>           [,1]      [,2]      [,3]      [,4]       [,5]      [,6]      [,7]
#> [1,] 0.7899726 0.3421906 0.0000000 0.0000000 0.08468105 0.0000000 0.0000000
#> [2,] 1.0001995 0.0000000 0.4570574 0.0000000 0.00000000 0.2120659 0.0000000
#> [3,] 1.0666814 0.0000000 0.0000000 0.5880327 0.00000000 0.0000000 0.2736337
BootCI(pb)
#> $ll
#>           [,1]      [,2]      [,3]        [,4]        [,5]     [,6]
#> [1,] 0.5557074 0.2348853 0.0000000 -0.01932668  0.00000000 0.000000
#> [2,] 0.9050943 0.0000000 0.3620924  0.00000000 -0.01125434 0.110489
#> [3,] 0.9168932 0.0000000 0.0000000  0.46891841  0.00000000 0.000000
#>              [,7]
#> [1,] -0.008895035
#> [2,]  0.000000000
#> [3,]  0.162752002
#> 
#> $ul
#>          [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#> [1,] 1.183012 0.3181934 0.0000000 0.0000000 0.1005727 0.0000000 0.0000000
#> [2,] 1.396692 0.0000000 0.4905112 0.0000000 0.0000000 0.2018012 0.0000000
#> [3,] 1.318093 0.0000000 0.0000000 0.6296753 0.0000000 0.0000000 0.3195748
#> 

# Residual bootstrap
system.time(
  rb <- RBootVAROLS(
    data = dat_p2,
    p = 2,
    B = 10
  )
)
#>    user  system elapsed 
#>   0.004   0.000   0.002 
rb$est
#>           [,1]        [,2]        [,3]        [,4]        [,5]         [,6]
#> [1,] 0.7899726  0.36836768 0.008539283  0.02231498  0.11338780  0.024505442
#> [2,] 1.0001995  0.01334496 0.481949237 -0.01730029 -0.03223315  0.234373552
#> [3,] 1.0666814 -0.03022951 0.014945306  0.60741736 -0.01610408 -0.006149189
#>             [,7]
#> [1,] -0.01143384
#> [2,]  0.01815779
#> [3,]  0.29620975
BootCI(rb)
#> $ll
#>           [,1]        [,2]        [,3]        [,4]        [,5]         [,6]
#> [1,] 0.4536865  0.33646242 -0.04504321 -0.04958349  0.07526484  0.001450026
#> [2,] 0.6743771 -0.05744434  0.43100317 -0.09000561 -0.08508199  0.217092195
#> [3,] 0.8285059 -0.06885820 -0.02988599  0.58366382 -0.06684358 -0.065341556
#>             [,7]
#> [1,] -0.06856902
#> [2,] -0.06012427
#> [3,]  0.23097354
#> 
#> $ul
#>           [,1]        [,2]       [,3]       [,4]        [,5]       [,6]
#> [1,] 0.9540287 0.388266044 0.03441555 0.06190496 0.165386209 0.08599289
#> [2,] 1.2081003 0.040543620 0.52770238 0.04936992 0.039412862 0.27919637
#> [3,] 1.3902837 0.006705037 0.05717504 0.66021985 0.002434246 0.04184917
#>            [,7]
#> [1,] 0.05479999
#> [2,] 0.09880127
#> [3,] 0.33775462
#> 
system.time(
  rb <- RBootVARLasso(
    data = dat_p2,
    p = 2,
    B = 10,
    n_lambdas = 100,
    crit = "ebic",
    max_iter = 1000,
    tol = 1e-5
  )
)
#>    user  system elapsed 
#>  24.315   5.783  15.170 
rb$est
#>           [,1]      [,2]      [,3]      [,4]       [,5]      [,6]      [,7]
#> [1,] 0.7899726 0.3421906 0.0000000 0.0000000 0.08468105 0.0000000 0.0000000
#> [2,] 1.0001995 0.0000000 0.4570574 0.0000000 0.00000000 0.2120659 0.0000000
#> [3,] 1.0666814 0.0000000 0.0000000 0.5880327 0.00000000 0.0000000 0.2736337
BootCI(rb)
#> $ll
#>           [,1]      [,2]      [,3]      [,4]       [,5]      [,6]     [,7]
#> [1,] 0.5975985 0.3247305 0.0000000 0.0000000 0.08036022 0.0000000 0.000000
#> [2,] 0.8720934 0.0000000 0.4317423 0.0000000 0.00000000 0.2003202 0.000000
#> [3,] 1.0603452 0.0000000 0.0000000 0.5600429 0.00000000 0.0000000 0.260609
#> 
#> $ul
#>          [,1]      [,2]      [,3]      [,4]       [,5]      [,6]      [,7]
#> [1,] 1.442011 0.3480685 0.0000000 0.0000000 0.08613561 0.0000000 0.0000000
#> [2,] 1.301586 0.0000000 0.4586674 0.0000000 0.00000000 0.2128129 0.0000000
#> [3,] 1.533059 0.0000000 0.0000000 0.5816201 0.00000000 0.0000000 0.2706496
#>