Fit Vector Autoregressive (VAR) Model using dynr on Each of the Data Matrix in a List

This function estimates the parameters of a VAR model using the dynr package for each of the data matrix in a list.

Usage

FitMLVARDynr(
  data,
  p,
  ncores = 1,
  conf_level = 0.95,
  optimization_flag = TRUE,
  hessian_flag = FALSE,
  verbose = FALSE,
  weight_flag = FALSE,
  debug_flag = FALSE,
  perturb_flag = FALSE
)

Arguments

data

List. Each element is a numeric matrix of time series data with dimensions t by k, where t is the number of observations and k is the number of variables.

p

Positive integer. Number of lags. Only supports p = 1 and p = 2.

ncores

Positive integer. Number of cores to use. Not supported on Windows.

conf_level

a cumulative proportion indicating the level of desired confidence intervals for the final parameter estimates (default is .95)

optimization_flag

a flag (TRUE/FALSE) indicating whether optimization is to be done.

hessian_flag

a flag (TRUE/FALSE) indicating whether the Hessian matrix is to be calculated.

verbose

a flag (TRUE/FALSE) indicating whether more detailed intermediate output during the estimation process should be printed

weight_flag

a flag (TRUE/FALSE) indicating whether the negative log likelihood function should be weighted by the length of the time series for each individual

debug_flag

a flag (TRUE/FALSE) indicating whether users want additional dynr output that can be used for diagnostic purposes

perturb_flag

a flag (TRUE/FLASE) indicating whether to perturb the latent states during estimation. Only useful for ensemble forecasting.

Value

A list each element of which is an object of class dynrCook.

See also

Author

Ivan Jacob Agaloos Pesigan