Parametric Bootstrap for the Vector Autoregressive Model Using Lasso Regularization

Parametric Bootstrap for the Vector Autoregressive Model Using Lasso Regularization

Usage

PBootVARLasso(data, p, B, burn_in, n_lambdas, crit, max_iter, tol)

Arguments

data

Numeric matrix. The time series data with dimensions t by k, where t is the number of observations and k is the number of variables.

p

Integer. The order of the VAR model (number of lags).

B

Integer. Number of bootstrap samples to generate.

burn_in

Integer. Number of burn-in observations to exclude before returning the results in the simulation step.

n_lambdas

Integer. Number of lambdas to generate.

crit

Character string. Information criteria to use. Valid values include "aic", "bic", and "ebic".

max_iter

Integer. The maximum number of iterations for the coordinate descent algorithm (e.g., max_iter = 10000).

tol

Numeric. Convergence tolerance. The algorithm stops when the change in coefficients between iterations is below this tolerance (e.g., tol = 1e-5).

Value

List with the following elements:

est: Numeric matrix. Original Lasso estimate of the coefficient matrix.

boot: Numeric matrix. Matrix of vectorized bootstrap estimates of the coefficient matrix.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

PBootVARLasso(
  data = dat_p2,
  p = 2,
  B = 10,
  burn_in = 20,
  n_lambdas = 100,
  crit = "ebic",
  max_iter = 1000,
  tol = 1e-5
)
#> $est
#>           [,1]      [,2]      [,3]      [,4]       [,5]      [,6]      [,7]
#> [1,] 0.7899726 0.3421906 0.0000000 0.0000000 0.08468105 0.0000000 0.0000000
#> [2,] 1.0001995 0.0000000 0.4570574 0.0000000 0.00000000 0.2120659 0.0000000
#> [3,] 1.0666814 0.0000000 0.0000000 0.5880327 0.00000000 0.0000000 0.2736337
#> 
#> $boot
#>            [,1]      [,2]      [,3]      [,4] [,5] [,6] [,7]      [,8] [,9]
#>  [1,] 0.4974354 0.9819246 1.1738369 0.2874211    0    0    0 0.4556572    0
#>  [2,] 1.0830455 0.9508295 1.0430896 0.2693074    0    0    0 0.4549626    0
#>  [3,] 0.6755464 1.1632164 1.3881747 0.3515290    0    0    0 0.4471707    0
#>  [4,] 0.7017151 0.9256836 0.8637963 0.3257534    0    0    0 0.3862832    0
#>  [5,] 1.0700657 0.9049231 0.6395171 0.2977275    0    0    0 0.3834785    0
#>  [6,] 0.7185333 1.2510253 0.9615066 0.3072187    0    0    0 0.4219737    0
#>  [7,] 0.9427448 0.7641247 1.2277857 0.2996346    0    0    0 0.5027709    0
#>  [8,] 0.9015404 0.7686409 0.9286662 0.2645686    0    0    0 0.4017719    0
#>  [9,] 0.8224481 0.9757187 1.1222308 0.2962442    0    0    0 0.4237595    0
#> [10,] 1.1491493 1.2429397 0.9404054 0.2695900    0    0    0 0.3907207    0
#>       [,10] [,11]     [,12]      [,13] [,14] [,15] [,16]     [,17] [,18] [,19]
#>  [1,]     0     0 0.5797954 0.05004264     0     0     0 0.1674078     0     0
#>  [2,]     0     0 0.5461432 0.05396555     0     0     0 0.1733104     0     0
#>  [3,]     0     0 0.5677546 0.04139328     0     0     0 0.1903936     0     0
#>  [4,]     0     0 0.5907915 0.00000000     0     0     0 0.1657802     0     0
#>  [5,]     0     0 0.5889982 0.00000000     0     0     0 0.2217680     0     0
#>  [6,]     0     0 0.5252407 0.11276545     0     0     0 0.1616266     0     0
#>  [7,]     0     0 0.5411149 0.02913935     0     0     0 0.1835144     0     0
#>  [8,]     0     0 0.5496261 0.02280715     0     0     0 0.1972798     0     0
#>  [9,]     0     0 0.5494509 0.05684365     0     0     0 0.1618567     0     0
#> [10,]     0     0 0.5003265 0.00000000     0     0     0 0.1611615     0     0
#>       [,20]     [,21]
#>  [1,]     0 0.2168407
#>  [2,]     0 0.2892539
#>  [3,]     0 0.2067496
#>  [4,]     0 0.2192326
#>  [5,]     0 0.2226161
#>  [6,]     0 0.2987349
#>  [7,]     0 0.2563188
#>  [8,]     0 0.2544362
#>  [9,]     0 0.2286832
#> [10,]     0 0.2856441
#>