Parametric Bootstrap for the Vector Autoregressive Model Using Ordinary Least Squares

Parametric Bootstrap for the Vector Autoregressive Model Using Ordinary Least Squares

Usage

PBootVAROLS(data, p, B, burn_in)

Arguments

data

Numeric matrix. The time series data with dimensions t by k, where t is the number of observations and k is the number of variables.

p

Integer. The order of the VAR model (number of lags).

B

Integer. Number of bootstrap samples to generate.

burn_in

Integer. Number of burn-in observations to exclude before returning the results in the simulation step.

Value

List with the following elements:

est: Numeric matrix. Original OLS estimate of the coefficient matrix.

boot: Numeric matrix. Matrix of vectorized bootstrap estimates of the coefficient matrix.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

PBootVAROLS(data = dat_p2, p = 2, B = 10, burn_in = 20)
#> $est
#>           [,1]        [,2]        [,3]        [,4]        [,5]         [,6]
#> [1,] 0.7899726  0.36836768 0.008539283  0.02231498  0.11338780  0.024505442
#> [2,] 1.0001995  0.01334496 0.481949237 -0.01730029 -0.03223315  0.234373552
#> [3,] 1.0666814 -0.03022951 0.014945306  0.60741736 -0.01610408 -0.006149189
#>             [,7]
#> [1,] -0.01143384
#> [2,]  0.01815779
#> [3,]  0.29620975
#> 
#> $boot
#>            [,1]      [,2]      [,3]      [,4]         [,5]         [,6]
#>  [1,] 1.0339067 0.8563461 1.2147178 0.3043989  0.043006225 -0.010584587
#>  [2,] 0.7767891 0.9395002 1.2402251 0.3455775  0.013749552  0.020763805
#>  [3,] 0.8858024 0.9792760 1.0857908 0.3049715  0.018038038  0.022664823
#>  [4,] 0.7319226 1.0382510 1.1037546 0.3243381 -0.002071004 -0.039740135
#>  [5,] 0.8147346 1.1341339 1.2768715 0.3518268  0.061060342 -0.040368895
#>  [6,] 0.4678511 0.8506688 0.8719993 0.3099302 -0.010660569 -0.062050797
#>  [7,] 0.9616407 1.1437992 1.3570124 0.3714781 -0.016588251 -0.045152826
#>  [8,] 1.1355471 0.8277787 1.1788657 0.3714424 -0.038722084  0.022296050
#>  [9,] 0.9031797 0.7800243 0.7027008 0.3888607  0.065252694 -0.009833522
#> [10,] 0.6798369 0.8807969 1.1350965 0.3517783  0.029388522 -0.003021074
#>               [,7]      [,8]          [,9]        [,10]        [,11]     [,12]
#>  [1,]  0.022983706 0.4832843  0.0017451310  0.039606544 -0.003114556 0.6044163
#>  [2,]  0.048922965 0.4490854 -0.0506293942  0.072927290 -0.055888222 0.6483062
#>  [3,] -0.004877112 0.4305571  0.0075327686 -0.022139366 -0.017943697 0.5936358
#>  [4,]  0.034854586 0.4603699  0.0072864475  0.030007877 -0.009876834 0.6355099
#>  [5,] -0.036311836 0.5024621  0.0259210136  0.091855510 -0.046385975 0.5903474
#>  [6,]  0.002728421 0.4581626 -0.0065678572  0.078209715 -0.016568271 0.6185438
#>  [7,] -0.006100489 0.4761503  0.0380562205  0.034836010  0.007490582 0.5560382
#>  [8,]  0.010774972 0.4797289 -0.0001599374 -0.006170664 -0.022486013 0.5809807
#>  [9,]  0.018557601 0.4862995  0.0616480219  0.050746273  0.045420604 0.6759926
#> [10,] -0.010393793 0.4569010  0.0056861590 -0.025421918 -0.006035087 0.6190037
#>            [,13]       [,14]       [,15]        [,16]     [,17]        [,18]
#>  [1,] 0.14043234 -0.07067758  0.02463394  0.006950869 0.2168832 -0.003606205
#>  [2,] 0.08723934 -0.02420388 -0.07519354 -0.021860219 0.2549828  0.018480637
#>  [3,] 0.10760860 -0.01492266 -0.03356821  0.002683927 0.2691897  0.026259870
#>  [4,] 0.13272737 -0.04803265 -0.03399444 -0.007973215 0.2696608 -0.007773317
#>  [5,] 0.17696951 -0.04998886 -0.03334827  0.026762357 0.2181054  0.018574448
#>  [6,] 0.17406956  0.01011412  0.02669844  0.047671890 0.2951913  0.016288878
#>  [7,] 0.09287045 -0.03487430 -0.01870644  0.018408304 0.2038712  0.009061413
#>  [8,] 0.08975206 -0.01821016 -0.02354668  0.012637309 0.2723124 -0.005411039
#>  [9,] 0.05734899 -0.04009763  0.03852136  0.034556289 0.2060461 -0.019890320
#> [10,] 0.11037886 -0.02994464 -0.09549397  0.041375854 0.2832123  0.016573332
#>              [,19]        [,20]     [,21]
#>  [1,] -0.043536684  0.021826643 0.2795363
#>  [2,] -0.050032562  0.065164938 0.2475600
#>  [3,]  0.045178889  0.023312646 0.2867258
#>  [4,] -0.004275699  0.009513111 0.2708710
#>  [5,] -0.074751449  0.026889772 0.2871549
#>  [6,] -0.038058186  0.014466244 0.2971471
#>  [7,] -0.028707282 -0.001998587 0.3181315
#>  [8,] -0.002889202  0.039762517 0.3083487
#>  [9,] -0.052211867 -0.023671714 0.2341877
#> [10,]  0.043895602  0.008941731 0.2809307
#>