Residual Bootstrap for the Vector Autoregressive Model Using Ordinary Least Squares

Residual Bootstrap for the Vector Autoregressive Model Using Ordinary Least Squares

Usage

RBootVAROLS(data, p, B)

Arguments

data

Numeric matrix. The time series data with dimensions t by k, where t is the number of observations and k is the number of variables.

p

Integer. The order of the VAR model (number of lags).

B

Integer. Number of bootstrap samples to generate.

Value

List with the following elements:

List of bootstrap estimates

original X

List of bootstrapped Y

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

rb <- RBootVAROLS(data = dat_p2, p = 2, B = 10)
str(rb)
#> List of 3
#>  $ coef:List of 10
#>   ..$ : num [1:3, 1:7] 0.8055 1.132 1.1283 0.3567 0.0102 ...
#>   ..$ : num [1:3, 1:7] 0.8917 1.135 1.0837 0.3704 0.0727 ...
#>   ..$ : num [1:3, 1:7] 0.6836 1.1486 1.1955 0.3488 0.0837 ...
#>   ..$ : num [1:3, 1:7] 0.8033 0.9775 1.4875 0.3618 0.0377 ...
#>   ..$ : num [1:3, 1:7] 1.001 0.781 1.227 0.339 -0.035 ...
#>   ..$ : num [1:3, 1:7] 0.9866 0.9285 1.2667 0.3215 0.0814 ...
#>   ..$ : num [1:3, 1:7] 0.539 0.8439 0.6571 0.422 -0.0153 ...
#>   ..$ : num [1:3, 1:7] 0.571 1.005 1.256 0.359 -0.016 ...
#>   ..$ : num [1:3, 1:7] 0.92899 1.27726 1.07878 0.41553 0.00888 ...
#>   ..$ : num [1:3, 1:7] 0.7275 1.1262 0.9611 0.3447 0.0302 ...
#>  $ X   : num [1:998, 1:7] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ Y   :List of 10
#>   ..$ : num [1:998, 1:3] 1.027 0.241 1.651 2.484 2.838 ...
#>   ..$ : num [1:998, 1:3] 0.241 0.965 0.423 1.949 0.436 ...
#>   ..$ : num [1:998, 1:3] 0.856 0.675 2.596 2.304 2.026 ...
#>   ..$ : num [1:998, 1:3] 2.23 1.94 3.02 1.41 1.46 ...
#>   ..$ : num [1:998, 1:3] 2.27 2.003 1.149 0.798 2.478 ...
#>   ..$ : num [1:998, 1:3] 1.077 1.948 0.653 2.351 1.483 ...
#>   ..$ : num [1:998, 1:3] -0.476 1.16 1.098 3.053 1.791 ...
#>   ..$ : num [1:998, 1:3] 3.315 1.124 -0.487 1.436 1.83 ...
#>   ..$ : num [1:998, 1:3] 0.683 -1.069 -0.163 1.295 1.629 ...
#>   ..$ : num [1:998, 1:3] 0.244 2.874 1.639 1.565 0.674 ...